+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## (2022-01-20)
+- added azure-ts-sqlserver
+
 ## (2021-12-16)
 - added aws-classic-ts-eks-awsx-spot